[984] Improve read parquet to include all columns
